// src/admin-actions/enums/admin-action-type.enum.ts
export enum AdminActionType {
  BANNED_WORD_CREATE = 'BANNED_WORD_CREATE',
  BANNED_WORD_DELETE = 'BANNED_WORD_DELETE',
  ADMIN_USER_LIST = 'ADMIN_USER_LIST',
  ADMIN_STAR_USER_LIST = 'ADMIN_STAR_USER_LIST',
  BANNED_NICK_CREATE = 'BANNED_NICK_CREATE',
  BANNED_NICK_DELETE = 'BANNED_NICK_DELETE',
  STATUS_MODE_CREATE = 'STATUS_MODE_CREATE',
  STATUS_MODE_UPDATE = 'STATUS_MODE_UPDATE',
  STATUS_MODE_DELETE = 'STATUS_MODE_DELETE',
  USER_STATUS_MODE_UPDATE = 'USER_STATUS_MODE_UPDATE',
  USER_FRAME_UPDATE = 'USER_FRAME_UPDATE',
  ADMIN_USER_UPDATE = 'ADMIN_USER_UPDATE',
  ADMIN_USER_DELETE = 'ADMIN_USER_DELETE',
  ADMIN_USER_BAN = 'ADMIN_USER_BAN',
  ADMIN_DEVICE_BAN = 'ADMIN_DEVICE_BAN',
  ADMIN_USER_UNBAN = 'ADMIN_USER_UNBAN',
  ADMIN_IP_UNBAN = 'ADMIN_IP_UNBAN',
  ADMIN_KICK_USER = 'ADMIN_KICK_USER',
  ADMIN_MIC_BAN_TOGGLE = 'ADMIN_MIC_BAN_TOGGLE',
  ADMIN_CAMERA_BAN_TOGGLE = 'ADMIN_CAMERA_BAN_TOGGLE',
  ADMIN_ROOM_MUTE_TOGGLE = 'ADMIN_ROOM_MUTE_TOGGLE',
  ADMIN_GLOBAL_MUTE_TOGGLE = 'ADMIN_GLOBAL_MUTE_TOGGLE',
  SYSTEM_RESET = 'SYSTEM_RESET',
}